refactor(network): split discovery into per-protocol modules
Convert the 2238-line src/network/discovery.rs into a
src/network/discovery/ directory of focused modules — one per
discovery protocol (mdns, tcp_connect, ssdp, udp_scan, arp, ws,
snmp_method, llmnr_method) plus coordinator/types/dns/privileges
helpers. Pure motion; the public surface (discover_hosts,
DiscoveryEvent, DiscoveryEventSink, DiscoveryMethod, DiscoveryOptions,
DiscoveryResult, check_privileges, get_available_methods) is
re-exported through the new mod.rs facade. No behavior change.
