--- src-tauri/src/network/discovery.rs
@@ -1,2238 +1,0 @@
-//! Host discovery methods for network scanning
-//!
-//! Provides userspace methods for discovering live hosts without requiring
-//! raw socket privileges:
-//! - TCP Connect (probes common ports via standard OS connect)
-//! - mDNS/Bonjour (discovers advertised services)
-//! - SSDP/UPnP (discovers UPnP devices)
-//! - UDP Scan (probes common UDP ports)
-//! - WS-Discovery (discovers Windows devices and printers)
-//! - ARP Cache (reads the OS ARP table)
-
-use std::collections::HashSet;
-use std::net::{IpAddr, SocketAddr};
-use std::sync::Arc;
-use std::time::Duration;
-
-use futures::stream::FuturesUnordered;
-use futures::StreamExt;
-use serde::{Deserialize, Serialize};
-use tokio::sync::Semaphore;
-use tokio::time::timeout;
-
-use super::scanner::ScanState;
-
-use super::interfaces::discover_mdns_services;
-
-/// Discovery method to use for finding live hosts
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum DiscoveryMethod {
-    /// TCP Connect scan to common ports (no privileges needed)
-    #[default]
-    TcpConnect,
-    /// mDNS/Bonjour service discovery
-    Mdns,
-    /// SSDP/UPnP discovery for smart devices
-    Ssdp,
-    /// UDP scan to common ports (DNS, NetBIOS, SNMP)
-    UdpScan,
-    /// WS-Discovery (Windows devices, printers)
-    WsDiscovery,
-    /// ARP cache reading (no privileges needed)
-    ArpCache,
-    /// SNMP sysName broadcast across the target range
-    Snmp,
-    /// LLMNR multicast PTR queries across the target range
-    Llmnr,
-    /// Skip host discovery, scan all targets
-    None,
-}
-
-impl std::fmt::Display for DiscoveryMethod {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let name = match self {
-            Self::TcpConnect => "tcp_connect",
-            Self::Mdns => "mdns",
-            Self::Ssdp => "ssdp",
-            Self::UdpScan => "udp_scan",
-            Self::WsDiscovery => "ws_discovery",
-            Self::ArpCache => "arp_cache",
-            Self::Snmp => "snmp",
-            Self::Llmnr => "llmnr",
-            Self::None => "none",
-        };
-        write!(f, "{name}")
-    }
-}
-
-/// Host metadata collected during discovery
-#[derive(Debug, Clone, Serialize, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct HostMetadata {
-    /// Hostname (from mDNS, DNS reverse lookup, `NetBIOS`, etc.)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub hostname: Option<String>,
-    /// Source of the hostname resolution (dns, mdns, netbios, snmp, tls)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub hostname_source: Option<String>,
-    /// `NetBIOS` name (from `NetBIOS` Node Status query)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub netbios_name: Option<String>,
-    /// MAC address (from ARP scan)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub mac_address: Option<String>,
-    /// Vendor name (from OUI lookup)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub vendor: Option<String>,
-    /// mDNS services advertised by this host
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub mdns_services: Vec<MdnsServiceInfo>,
-    /// SSDP/UPnP device information
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ssdp_device: Option<super::types::SsdpDeviceInfo>,
-    /// WS-Discovery device information
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ws_discovery: Option<super::ws_discovery::WsDiscoveryInfo>,
-    /// SNMP device information (sysName, sysDescr, etc.)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub snmp_info: Option<super::snmp::SnmpDeviceInfo>,
-    /// TLS certificate Subject Alternative Names (dNSName)
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub tls_names: Vec<String>,
-    /// Service banners collected via TCP banner-grab
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub banners: Vec<super::banner::ServiceBanner>,
-}
-
-/// mDNS service information for a host
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct MdnsServiceInfo {
-    /// Service instance name
-    pub instance_name: String,
-    /// Service type (e.g., "_http._tcp")
-    pub service_type: String,
-    /// Port number
-    pub port: u16,
-    /// TXT record properties as key-value pairs
-    pub properties: Vec<(String, String)>,
-}
-
-/// Result of host discovery
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DiscoveryResult {
-    /// Method used for discovery
-    pub method: String,
-    /// Discovered hosts (IP addresses)
-    pub hosts: Vec<String>,
-    /// Hostname mapping: IP address -> hostname (for mDNS discovery)
-    /// Deprecated: Use `host_metadata` instead
-    pub hostnames: std::collections::HashMap<String, String>,
-    /// Extended host metadata (MAC addresses, vendors, mDNS services)
-    pub host_metadata: std::collections::HashMap<String, HostMetadata>,
-    /// Hosts that failed discovery or timed out
-    pub unreachable: Vec<String>,
-    /// Duration of discovery in milliseconds
-    pub duration_ms: u64,
-    /// Error message if discovery failed
-    pub error: Option<String>,
-    /// Whether the method requires elevated privileges
-    pub requires_privileges: bool,
-}
-
-/// Discovery options
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DiscoveryOptions {
-    /// Methods to use (in order of preference)
-    pub methods: Vec<DiscoveryMethod>,
-    /// Timeout per host in milliseconds
-    pub timeout_ms: u32,
-    /// Maximum concurrent discovery operations
-    pub concurrency: u32,
-    /// Ports to probe for TCP SYN scan
-    pub syn_ports: Option<Vec<u16>>,
-    /// mDNS service types to discover
-    pub mdns_services: Option<Vec<String>>,
-    /// Whether to resolve `NetBIOS` names for discovered hosts
-    #[serde(default)]
-    pub resolve_netbios: bool,
-}
-
-impl Default for DiscoveryOptions {
-    fn default() -> Self {
-        Self {
-            methods: vec![DiscoveryMethod::TcpConnect],
-            timeout_ms: 1000,
-            concurrency: 100,
-            syn_ports: Some(vec![80, 443, 22, 445, 139]),
-            mdns_services: None,
-            resolve_netbios: false,
-        }
-    }
-}
-
-/// Streaming event sent to the frontend via Tauri Channel API
-///
-/// Uses discriminated union serialization for type-safe handling in TypeScript.
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase", tag = "event", content = "data")]
-pub enum DiscoveryEvent {
-    /// A discovery method has started execution
-    MethodStarted {
-        /// Name of the discovery method
-        method: String,
-    },
-    /// A discovery method completed — hosts are immediately available for display
-    MethodCompleted {
-        /// The discovery result for this method
-        result: DiscoveryResult,
-    },
-    /// Hostname resolution phase has started
-    ResolvingHostnames,
-    /// All discovery and resolution completed — final merged results
-    Completed {
-        /// Final merged results from all methods
-        results: Vec<DiscoveryResult>,
-    },
-    /// Discovery was cancelled by the user
-    Cancelled,
-}
-
-/// Generic event sink for streaming discovery results (Tauri-independent)
-///
-/// Implementations:
-/// - Main process: forwards events to Tauri `Channel<DiscoveryEvent>`
-/// - Sidecar: writes JSON Lines to stdout
-pub trait DiscoveryEventSink: Send + Sync {
-    /// Send a discovery event to the consumer
-    fn send(&self, event: DiscoveryEvent) -> Result<(), String>;
-}
-
-/// Extended ports for TCP connect discovery (includes IoT and common services)
-pub const CONNECT_DISCOVERY_PORTS: &[u16] = &[
-    21,    // FTP
-    22,    // SSH
-    23,    // Telnet
-    25,    // SMTP
-    53,    // DNS
-    80,    // HTTP
-    110,   // POP3
-    135,   // MS RPC
-    139,   // NetBIOS
-    143,   // IMAP
-    443,   // HTTPS
-    445,   // SMB
-    515,   // LPD (printers)
-    548,   // AFP (Apple Filing Protocol)
-    554,   // RTSP (cameras)
-    631,   // IPP (CUPS printing)
-    993,   // IMAPS
-    995,   // POP3S
-    1433,  // MS SQL
-    1883,  // MQTT (IoT)
-    3306,  // MySQL
-    3389,  // RDP
-    5000,  // UPnP
-    5001,  // Synology DSM
-    5353,  // mDNS
-    5900,  // VNC
-    6379,  // Redis
-    8000,  // HTTP Alt
-    8008,  // Chromecast
-    8080,  // HTTP Alt
-    8081,  // HTTP Alt
-    8443,  // HTTPS Alt
-    8888,  // HTTP Alt
-    9000,  // Various
-    9100,  // JetDirect (printers)
-    49152, // UPnP / Dynamic ports
-    62078, // iPhone sync
-];
-
-/// Discover live hosts using specified methods in parallel with streaming results
-///
-/// Discovery methods run concurrently via `FuturesUnordered`, streaming each method's
-/// result to the frontend via the Tauri Channel API as soon as it completes.
-/// Supports cancellation via `ScanState`. Local IP addresses are automatically included.
-pub async fn discover_hosts(
-    targets: &[IpAddr],
-    options: &DiscoveryOptions,
-    on_event: &dyn DiscoveryEventSink,
-    cancel_state: &Arc<ScanState>,
-) -> Vec<DiscoveryResult> {
-    let methods = options.methods.clone();
-
-    // Get local IP addresses to auto-include them in results
-    let local_ips = get_local_ip_addresses();
-    let local_ips_in_targets: Vec<String> = targets
-        .iter()
-        .filter(|ip| local_ips.contains(ip))
-        .map(ToString::to_string)
-        .collect();
-
-    // Separate ARP cache from other methods for deferred execution.
-    // Running ARP cache after other methods improves detection: packets sent during
-    // Phase 1 (ICMP, TCP Connect, etc.) populate the OS ARP table, which Phase 2 reads.
-    let has_arp_cache = methods.contains(&DiscoveryMethod::ArpCache);
-    let phase1_methods: Vec<DiscoveryMethod> = methods
-        .iter()
-        .filter(|m| **m != DiscoveryMethod::ArpCache)
-        .copied()
-        .collect();
-
-    // Phase 1: Execute non-ARP-cache methods in parallel, streaming results
-    let mut results =
-        run_discovery_methods_streaming(&phase1_methods, targets, options, on_event, cancel_state)
-            .await;
-
-    // Check cancellation after Phase 1
-    if cancel_state.is_cancelled() {
-        return results;
-    }
-
-    // Phase 2: Execute ARP cache after other methods complete
-    // Brief delay allows OS ARP table to be populated by Phase 1 packets
-    if has_arp_cache {
-        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-        let arp_results = run_discovery_methods_streaming(
-            &[DiscoveryMethod::ArpCache],
-            targets,
-            options,
-            on_event,
-            cancel_state,
-        )
-        .await;
-        results.extend(arp_results);
-
-        if cancel_state.is_cancelled() {
-            return results;
-        }
-    }
-
-    // Add local IP addresses to results (they are always "alive")
-    if !local_ips_in_targets.is_empty() {
-        // Add local IPs to every successful discovery result
-        for result in &mut results {
-            if result.error.is_none() {
-                for local_ip in &local_ips_in_targets {
-                    if !result.hosts.contains(local_ip) {
-                        result.hosts.push(local_ip.clone());
-                    }
-                    // Remove from unreachable if present
-                    result.unreachable.retain(|ip| ip != local_ip);
-                    // Add metadata marking it as local
-                    result.host_metadata.entry(local_ip.clone()).or_default();
-                }
-            }
-        }
-
-        // Also add a special "local" discovery result for tracking
-        let local_result = DiscoveryResult {
-            method: "local".to_string(),
-            hosts: local_ips_in_targets.clone(),
-            hostnames: std::collections::HashMap::new(),
-            host_metadata: local_ips_in_targets
-                .iter()
-                .map(|ip| (ip.clone(), HostMetadata::default()))
-                .collect(),
-            unreachable: vec![],
-            duration_ms: 0,
-            error: None,
-            requires_privileges: false,
-        };
-        let _ = on_event.send(DiscoveryEvent::MethodCompleted {
-            result: local_result.clone(),
-        });
-        results.push(local_result);
-    }
-
-    // Resolve hostnames for all discovered hosts if enabled
-    if options.resolve_netbios {
-        let _ = on_event.send(DiscoveryEvent::ResolvingHostnames);
-        resolve_hostnames_for_results(&mut results, options).await;
-    }
-
-    // Send final completed event with fully resolved results
-    let _ = on_event.send(DiscoveryEvent::Completed {
-        results: results.clone(),
-    });
-
-    results
-}
-
-/// Execute a single discovery method
-async fn execute_discovery_method(
-    method: DiscoveryMethod,
-    targets: &[IpAddr],
-    options: &DiscoveryOptions,
-) -> DiscoveryResult {
-    match method {
-        DiscoveryMethod::TcpConnect => tcp_connect_discovery(targets, options).await,
-        DiscoveryMethod::Mdns => mdns_discovery(options).await,
-        DiscoveryMethod::Ssdp => ssdp_discovery(options).await,
-        DiscoveryMethod::UdpScan => udp_scan_discovery(targets, options).await,
-        DiscoveryMethod::WsDiscovery => ws_discovery_method(options).await,
-        DiscoveryMethod::ArpCache => arp_cache_discovery(targets),
-        DiscoveryMethod::Snmp => snmp_discovery(targets, options).await,
-        DiscoveryMethod::Llmnr => llmnr_discovery(targets, options).await,
-        DiscoveryMethod::None => DiscoveryResult {
-            method: "none".to_string(),
-            hosts: targets.iter().map(ToString::to_string).collect(),
-            hostnames: std::collections::HashMap::new(),
-            host_metadata: std::collections::HashMap::new(),
-            unreachable: vec![],
-            duration_ms: 0,
-            error: None,
-            requires_privileges: false,
-        },
-    }
-}
-
-/// Execute multiple discovery methods in parallel, streaming results as each completes
-///
-/// Uses `FuturesUnordered` instead of `join_all` to emit each method's result to the
-/// frontend via Channel as soon as it completes. Supports cancellation between method
-/// completions via `ScanState`.
-async fn run_discovery_methods_streaming(
-    methods: &[DiscoveryMethod],
-    targets: &[IpAddr],
-    options: &DiscoveryOptions,
-    on_event: &dyn DiscoveryEventSink,
-    cancel_state: &Arc<ScanState>,
-) -> Vec<DiscoveryResult> {
-    if methods.is_empty() {
-        return vec![];
-    }
-
-    let mut futs: FuturesUnordered<_> = methods
-        .iter()
-        .map(|method| {
-            let method = *method;
-            let method_name = method.to_string();
-            async move {
-                // Notify frontend that this method has started
-                let _ = on_event.send(DiscoveryEvent::MethodStarted {
-                    method: method_name,
-                });
-
-                execute_discovery_method(method, targets, options).await
-            }
-        })
-        .collect();
-
-    let mut results = Vec::with_capacity(methods.len());
-
-    while let Some(result) = futs.next().await {
-        // Check cancellation between method completions
-        if cancel_state.is_cancelled() {
-            let _ = on_event.send(DiscoveryEvent::Cancelled);
-            break;
-        }
-
-        // Stream the completed result to the frontend
-        let _ = on_event.send(DiscoveryEvent::MethodCompleted {
-            result: result.clone(),
-        });
-        results.push(result);
-    }
-
-    results
-}
-
-/// Resolve hostnames for all discovered hosts via DNS PTR, `NetBIOS`, and LLMNR
-///
-/// Resolution priority (highest to lowest):
-/// 1. mDNS (already resolved during discovery)
-/// 2. DNS PTR (reverse lookup)
-/// 3. NetBIOS (Windows hosts)
-/// 4. LLMNR (Link-Local Multicast Name Resolution)
-#[allow(clippy::cognitive_complexity, clippy::too_many_lines)]
-async fn resolve_hostnames_for_results(
-    results: &mut [DiscoveryResult],
-    options: &DiscoveryOptions,
-) {
-    use super::netbios::resolve_netbios_name;
-
-    let timeout_duration = Duration::from_millis(u64::from(options.timeout_ms));
-
-    // Collect all unique discovered IPs
-    let all_ips_set: HashSet<IpAddr> = results
-        .iter()
-        .flat_map(|r| r.hosts.iter())
-        .filter_map(|h| h.parse::<IpAddr>().ok())
-        .collect();
-    let all_ips: Vec<IpAddr> = all_ips_set.iter().copied().collect();
-
-    // Phase 1: DNS PTR reverse lookup (fastest, works on corporate networks)
-    // Use shorter timeout for DNS as it's typically fast
-    let dns_timeout = Duration::from_millis(u64::from(options.timeout_ms).min(3000));
-    let dns_ptr_names =
-        resolve_dns_ptr_batch(&all_ips, dns_timeout, options.concurrency as usize).await;
-
-    // Apply DNS PTR results to metadata (only if no hostname exists yet)
-    for result in results.iter_mut() {
-        for host_ip in &result.hosts {
-            if let Some(dns_name) = dns_ptr_names.get(host_ip) {
-                let metadata = result.host_metadata.entry(host_ip.clone()).or_default();
-
-                // Only set if no hostname exists (mDNS has higher priority)
-                if metadata.hostname.is_none() {
-                    metadata.hostname = Some(dns_name.clone());
-                    metadata.hostname_source = Some("dns".to_string());
-                }
-            }
-        }
-    }
-
-    // Phase 2: Query SNMP device info concurrently (sysName, sysDescr, etc.)
-    // Use shorter timeout for SNMP as it should respond quickly or not at all
-    let snmp_timeout = Duration::from_millis(u64::from(options.timeout_ms).min(2000));
-    let snmp_sem = Arc::new(Semaphore::new(options.concurrency as usize));
-    let mut snmp_handles = Vec::new();
-
-    for &ip in &all_ips {
-        let sem = Arc::clone(&snmp_sem);
-        let timeout = snmp_timeout;
-
-        snmp_handles.push(tokio::spawn(async move {
-            let _permit = sem.acquire().await.ok()?;
-            let info = super::snmp::query_snmp_device_info(ip, timeout).await?;
-            Some((ip.to_string(), info))
-        }));
-    }
-
-    // Collect SNMP results
-    let mut snmp_info_map: std::collections::HashMap<String, super::snmp::SnmpDeviceInfo> =
-        std::collections::HashMap::new();
-    for handle in snmp_handles {
-        if let Ok(Some((ip, info))) = handle.await {
-            snmp_info_map.insert(ip, info);
-        }
-    }
-
-    // Apply SNMP results to metadata
-    for result in results.iter_mut() {
-        for host_ip in &result.hosts {
-            if let Some(snmp_info) = snmp_info_map.remove(host_ip) {
-                let metadata = result.host_metadata.entry(host_ip.clone()).or_default();
-
-                // Use sysName as hostname if no hostname exists yet
-                if metadata.hostname.is_none() {
-                    if let Some(ref sys_name) = snmp_info.sys_name {
-                        metadata.hostname = Some(sys_name.clone());
-                        metadata.hostname_source = Some("snmp".to_string());
-                    }
-                }
-
-                metadata.snmp_info = Some(snmp_info);
-            }
-        }
-    }
-
-    // Phase 3: Extract TLS certificate SAN names (HTTPS hosts)
-    // Use short timeout as we only probe port 443
-    let tls_timeout = Duration::from_millis(u64::from(options.timeout_ms).min(2000));
-    let tls_sem = Arc::new(Semaphore::new(options.concurrency as usize));
-    let mut tls_handles = Vec::new();
-
-    for &ip in &all_ips {
-        let sem = Arc::clone(&tls_sem);
-        let timeout = tls_timeout;
-
-        tls_handles.push(tokio::spawn(async move {
-            let _permit = sem.acquire().await.ok()?;
-
-            // Only probe port 443 for TLS
-            let names = super::tls_info::extract_tls_san_names(ip, 443, timeout).await;
-
-            if names.is_empty() {
-                None
-            } else {
-                Some((ip.to_string(), names))
-            }
-        }));
-    }
-
-    // Collect TLS SAN results
-    let mut tls_names_map: std::collections::HashMap<String, Vec<String>> =
-        std::collections::HashMap::new();
-    for handle in tls_handles {
-        if let Ok(Some((ip, names))) = handle.await {
-            tls_names_map.insert(ip, names);
-        }
-    }
-
-    // Apply TLS SAN results to metadata
-    for result in results.iter_mut() {
-        for host_ip in &result.hosts {
-            if let Some(tls_names) = tls_names_map.remove(host_ip) {
-                let metadata = result.host_metadata.entry(host_ip.clone()).or_default();
-
-                // Use first TLS SAN as hostname if no hostname exists yet
-                if metadata.hostname.is_none() && !tls_names.is_empty() {
-                    metadata.hostname = Some(tls_names[0].clone());
-                    metadata.hostname_source = Some("tls".to_string());
-                }
-
-                metadata.tls_names = tls_names;
-            }
-        }
-    }
-
-    // Phase 4: Resolve NetBIOS names concurrently (Windows hosts)
-    let semaphore = Arc::new(Semaphore::new(options.concurrency as usize));
-    let mut handles: Vec<tokio::task::JoinHandle<Option<(String, String)>>> = Vec::new();
-
-    for ip in all_ips {
-        let sem = Arc::clone(&semaphore);
-        let timeout = timeout_duration;
-
-        let handle: tokio::task::JoinHandle<Option<(String, String)>> = tokio::spawn(async move {
-            let _permit = sem.acquire().await.ok()?;
-
-            // NetBIOS resolution is blocking, run in a blocking task
-            let result = tokio::task::spawn_blocking(move || resolve_netbios_name(ip, timeout))
-                .await
-                .ok()
-                .flatten();
-
-            result.map(|name| (ip.to_string(), name))
-        });
-
-        handles.push(handle);
-    }
-
-    // Collect results
-    let mut netbios_names: std::collections::HashMap<String, String> =
-        std::collections::HashMap::new();
-    for handle in handles {
-        if let Ok(Some((ip, name))) = handle.await {
-            netbios_names.insert(ip, name);
-        }
-    }
-
-    // Phase 5: Try LLMNR for IPs without a hostname (fallback resolution)
-    let llmnr_timeout = Duration::from_millis(u64::from(options.timeout_ms).min(2000));
-    let mut llmnr_names: std::collections::HashMap<String, String> =
-        std::collections::HashMap::new();
-
-    // Collect IPs that don't have a hostname from any discovery method or DNS PTR
-    let ips_without_hostname: Vec<String> = {
-        let mut has_hostname: HashSet<String> = HashSet::new();
-        for result in results.iter() {
-            // Check mDNS hostnames
-            for ip in result.hostnames.keys() {
-                has_hostname.insert(ip.clone());
-            }
-            // Check host_metadata for DNS PTR resolved hostnames
-            for (ip, metadata) in &result.host_metadata {
-                if metadata.hostname.is_some() {
-                    has_hostname.insert(ip.clone());
-                }
-            }
-        }
-        all_ips_set
-            .iter()
-            .map(ToString::to_string)
-            .filter(|ip| !has_hostname.contains(ip))
-            .collect()
-    };
-
-    let llmnr_sem = Arc::new(Semaphore::new(options.concurrency as usize));
-    let mut llmnr_handles = Vec::new();
-
-    for ip_str in ips_without_hostname {
-        let sem = Arc::clone(&llmnr_sem);
-        let llmnr_t = llmnr_timeout;
-
-        llmnr_handles.push(tokio::spawn(async move {
-            let _permit = sem.acquire().await.ok()?;
-            let name = super::llmnr::resolve_hostname(&ip_str, llmnr_t).await?;
-            Some((ip_str, name))
-        }));
-    }
-
-    for handle in llmnr_handles {
-        if let Ok(Some((ip, name))) = handle.await {
-            llmnr_names.insert(ip, name);
-        }
-    }
-
-    // Update host_metadata in all results with NetBIOS and LLMNR names
-    for result in results.iter_mut() {
-        for host_ip in &result.hosts {
-            if let Some(netbios_name) = netbios_names.get(host_ip) {
-                let metadata = result
-                    .host_metadata
-                    .entry(host_ip.clone())
-                    .or_insert_with(HostMetadata::default);
-
-                metadata.netbios_name = Some(netbios_name.clone());
-
-                // Note: We intentionally do NOT add NetBIOS name to hostnames map.
-                // This is because NetBIOS names may differ from mDNS/DNS hostnames,
-                // and adding them would prevent proper host merging when the same
-                // host has both IPv4 (with NetBIOS) and IPv6 (with mDNS) addresses.
-                // The NetBIOS name is preserved in metadata.netbios_name for display.
-            }
-
-            // Add LLMNR hostname if found and no hostname exists yet
-            if let Some(llmnr_name) = llmnr_names.get(host_ip) {
-                let metadata = result
-                    .host_metadata
-                    .entry(host_ip.clone())
-                    .or_insert_with(HostMetadata::default);
-
-                if metadata.hostname.is_none() {
-                    metadata.hostname = Some(llmnr_name.clone());
-                    metadata.hostname_source = Some("llmnr".to_string());
-                }
-            }
-        }
-    }
-}
-
-// =============================================================================
-// DNS PTR Reverse Lookup
-// =============================================================================
-
-/// Resolve hostname via DNS PTR (reverse lookup) for a single IP
-async fn resolve_dns_ptr(ip: IpAddr) -> Option<String> {
-    use hickory_resolver::proto::rr::RData;
-    use hickory_resolver::Resolver;
-
-    // Create resolver with system configuration
-    let resolver = Resolver::builder_tokio().ok()?.build().ok()?;
-
-    // Perform reverse lookup
-    let response = resolver.reverse_lookup(ip).await.ok()?;
-
-    // Get the first PTR record and clean up the hostname
-    response.answers().iter().find_map(|record| {
-        if let RData::PTR(ptr) = &record.data {
-            let hostname = ptr.to_string();
-            Some(hostname.strip_suffix('.').unwrap_or(&hostname).to_string())
-        } else {
-            None
-        }
-    })
-}
-
-/// Resolve DNS PTR names for multiple IPs concurrently
-async fn resolve_dns_ptr_batch(
-    ips: &[IpAddr],
-    timeout_duration: Duration,
-    concurrency: usize,
-) -> std::collections::HashMap<String, String> {
-    let semaphore = Arc::new(Semaphore::new(concurrency));
-    let mut handles = Vec::new();
-
-    for &ip in ips {
-        let sem = Arc::clone(&semaphore);
-        let timeout_dur = timeout_duration;
-
-        handles.push(tokio::spawn(async move {
-            let _permit = sem.acquire().await.ok()?;
-
-            // Apply timeout to DNS lookup
-            let result = timeout(timeout_dur, resolve_dns_ptr(ip)).await.ok()?;
-
-            result.map(|name| (ip.to_string(), name))
-        }));
-    }
-
-    let mut results = std::collections::HashMap::new();
-    for handle in handles {
-        if let Ok(Some((ip, name))) = handle.await {
-            results.insert(ip, name);
-        }
-    }
-
-    results
-}
-
-// =============================================================================
-// mDNS Discovery
-// =============================================================================
-
-/// Minimum timeout for mDNS discovery (mDNS needs time for service resolution)
-const MDNS_MIN_TIMEOUT_MS: u32 = 5000;
-
-/// Common mDNS/Bonjour service types for host discovery
-///
-/// These service types are based on:
-/// - RFC 6763 (DNS-Based Service Discovery)
-/// - IANA Service Name and Transport Protocol Port Number Registry
-/// - Apple Bonjour Printing Specification
-/// - Common IoT and smart home protocols
-const DEFAULT_MDNS_SERVICE_TYPES: &[&str] = &[
-    // Web Services
-    "_http._tcp",  // HTTP web servers
-    "_https._tcp", // HTTPS web servers
-    // Remote Access
-    "_ssh._tcp",      // SSH servers
-    "_sftp-ssh._tcp", // SFTP over SSH
-    "_rfb._tcp",      // VNC (Remote Frame Buffer)
-    // File Sharing
-    "_smb._tcp",        // SMB/CIFS (Windows file sharing)
-    "_afpovertcp._tcp", // AFP (Apple Filing Protocol)
-    "_ftp._tcp",        // FTP servers
-    "_nfs._tcp",        // NFS (Network File System)
-    "_webdav._tcp",     // WebDAV
-    // Printing
-    "_ipp._tcp",            // Internet Printing Protocol
-    "_ipps._tcp",           // IPP over HTTPS
-    "_printer._tcp",        // LPR printers
-    "_pdl-datastream._tcp", // PDL Data Stream printers
-    "_scanner._tcp",        // Scanners
-    // Media & Streaming
-    "_airplay._tcp",         // Apple AirPlay
-    "_raop._tcp",            // Remote Audio Output Protocol (AirPlay audio)
-    "_daap._tcp",            // iTunes/DAAP (Digital Audio Access Protocol)
-    "_googlecast._tcp",      // Google Cast / Chromecast
-    "_spotify-connect._tcp", // Spotify Connect
-    // Smart Home / IoT
-    "_hap._tcp",     // HomeKit Accessory Protocol
-    "_homekit._tcp", // HomeKit
-    "_hue._tcp",     // Philips Hue
-    // Apple Services
-    "_device-info._tcp",    // Device information
-    "_companion-link._tcp", // Apple Watch companion
-    "_sleep-proxy._udp",    // Sleep Proxy (Wake-on-LAN)
-    "_adisk._tcp",          // Time Machine / Apple Disk
-    // Workstation Discovery
-    "_workstation._tcp", // macOS/Linux workstations
-    "_presence._tcp",    // Presence/availability
-    // Database Services
-    "_postgresql._tcp", // PostgreSQL
-    "_mysql._tcp",      // MySQL
-    // NAS Devices
-    "_nas._tcp",   // Generic NAS
-    "_iscsi._tcp", // iSCSI targets
-];
-
-/// Discover hosts using mDNS/Bonjour
-async fn mdns_discovery(options: &DiscoveryOptions) -> DiscoveryResult {
-    let start = std::time::Instant::now();
-
-    let service_types = options.mdns_services.clone().unwrap_or_else(|| {
-        DEFAULT_MDNS_SERVICE_TYPES
-            .iter()
-            .map(|s| (*s).to_string())
-            .collect()
-    });
-
-    // mDNS requires longer timeout for service discovery and resolution
-    let mdns_timeout = options.timeout_ms.max(MDNS_MIN_TIMEOUT_MS);
-
-    match discover_mdns_services(service_types, mdns_timeout).await {
-        Ok(results) => {
-            let mut hosts = HashSet::new();
-            let mut hostnames = std::collections::HashMap::new();
-            let mut host_metadata: std::collections::HashMap<String, HostMetadata> =
-                std::collections::HashMap::new();
-
-            for service in &results.services {
-                // Create mDNS service info
-                let service_info = MdnsServiceInfo {
-                    instance_name: service.instance_name.clone(),
-                    service_type: service.service_type.clone(),
-                    port: service.port,
-                    properties: service.properties.clone(),
-                };
-
-                for addr in &service.addresses {
-                    hosts.insert(addr.clone());
-
-                    // Map IP address to hostname (remove trailing dot if present)
-                    let hostname = service.hostname.trim_end_matches('.').to_string();
-                    if !hostname.is_empty() {
-                        hostnames.insert(addr.clone(), hostname.clone());
-                    }
-
-                    // Add or update host metadata
-                    host_metadata
-                        .entry(addr.clone())
-                        .and_modify(|meta| {
-                            // Update hostname if not set
-                            if meta.hostname.is_none() && !hostname.is_empty() {
-                                meta.hostname = Some(hostname.clone());
-                            }
-                            // Add service info
-                            meta.mdns_services.push(service_info.clone());
-                        })
-                        .or_insert_with(|| HostMetadata {
-                            hostname: if hostname.is_empty() {
-                                None
-                            } else {
-                                Some(hostname.clone())
-                            },
-                            hostname_source: if hostname.is_empty() {
-                                None
-                            } else {
-                                Some("mdns".to_string())
-                            },
-                            mdns_services: vec![service_info.clone()],
-                            ..Default::default()
-                        });
-                }
-            }
-
-            DiscoveryResult {
-                method: "mdns".to_string(),
-                hosts: hosts.into_iter().collect(),
-                hostnames,
-                host_metadata,
-                unreachable: vec![],
-                duration_ms: start.elapsed().as_millis() as u64,
-                error: None,
-                requires_privileges: false,
-            }
-        }
-        Err(e) => DiscoveryResult {
-            method: "mdns".to_string(),
-            hosts: vec![],
-            hostnames: std::collections::HashMap::new(),
-            host_metadata: std::collections::HashMap::new(),
-            unreachable: vec![],
-            duration_ms: start.elapsed().as_millis() as u64,
-            error: Some(e),
-            requires_privileges: false,
-        },
-    }
-}
-
-// =============================================================================
-// TCP Connect Discovery (Port Scan)
-// =============================================================================
-
-/// Discover hosts using TCP connect scan (no privileges needed)
-///
-/// After the first successful connect to any probe port, attempts a banner-grab
-/// against the recognised banner ports already in the candidate list and
-/// attaches the resulting [`ServiceBanner`]s to the host metadata.
-async fn tcp_connect_discovery(targets: &[IpAddr], options: &DiscoveryOptions) -> DiscoveryResult {
-    let start = std::time::Instant::now();
-    let timeout_duration = Duration::from_millis(u64::from(options.timeout_ms));
-    let semaphore = Arc::new(Semaphore::new(options.concurrency as usize));
-
-    // Use extended port list for better discovery accuracy
-    let ports = options
-        .syn_ports
-        .as_ref()
-        .map_or(CONNECT_DISCOVERY_PORTS, Vec::as_slice);
-
-    let handles: Vec<_> = targets
-        .iter()
-        .copied()
-        .map(|target| {
-            let sem = Arc::clone(&semaphore);
-            let ports = ports.to_vec();
-            tokio::spawn(async move {
-                let _permit = sem.acquire().await;
-                probe_target(target, &ports, timeout_duration).await
-            })
-        })
-        .collect();
-
-    let mut hosts = Vec::new();
-    let mut unreachable = Vec::new();
-    let mut host_metadata: std::collections::HashMap<String, HostMetadata> =
-        std::collections::HashMap::new();
-
-    for handle in handles {
-        if let Ok(probe) = handle.await {
-            let ip_str = probe.target.to_string();
-            if probe.reachable {
-                hosts.push(ip_str.clone());
-                if !probe.banners.is_empty() {
-                    host_metadata
-                        .entry(ip_str)
-                        .or_default()
-                        .banners
-                        .extend(probe.banners);
-                }
-            } else {
-                unreachable.push(ip_str);
-            }
-        }
-    }
-
-    DiscoveryResult {
-        method: "tcp_connect".to_string(),
-        hosts,
-        hostnames: std::collections::HashMap::new(),
-        host_metadata,
-        unreachable,
-        duration_ms: start.elapsed().as_millis() as u64,
-        error: None,
-        requires_privileges: false,
-    }
-}
-
-/// Result of probing a single target during TCP connect discovery.
-struct TcpProbeOutcome {
-    target: IpAddr,
-    reachable: bool,
-    banners: Vec<super::banner::ServiceBanner>,
-}
-
-/// Ports for which we know how to parse a banner. Limits banner-grab cost.
-const BANNER_GRAB_PORTS: &[u16] = &[21, 22, 23, 25, 80, 443, 587, 6379, 8080, 8443, 11211];
-
-/// Connect to each probe port and gather banners from supported services.
-async fn probe_target(
-    target: IpAddr,
-    ports: &[u16],
-    timeout_duration: Duration,
-) -> TcpProbeOutcome {
-    let mut reachable = false;
-    let mut banners = Vec::new();
-
-    for port in ports {
-        let addr = SocketAddr::new(target, *port);
-        let connect = timeout(timeout_duration, tokio::net::TcpStream::connect(addr)).await;
-        let stream = match connect {
-            Ok(Ok(stream)) => stream,
-            _ => continue,
-        };
-        reachable = true;
-
-        if BANNER_GRAB_PORTS.contains(port) {
-            if let Some(banner) = super::banner::grab_banner(stream, *port).await {
-                if !banners
-                    .iter()
-                    .any(|existing: &super::banner::ServiceBanner| existing == &banner)
-                {
-                    banners.push(banner);
-                }
-            }
-        }
-    }
-
-    TcpProbeOutcome {
-        target,
-        reachable,
-        banners,
-    }
-}
-
-// =============================================================================
-// Utility Functions
-// =============================================================================
-
-/// Check if the current process can run a given discovery method.
-///
-/// All remaining methods are userspace and require no special privileges,
-/// so this always returns `true`. Kept for API stability.
-pub fn check_privileges(_method: DiscoveryMethod) -> bool {
-    true
-}
-
-/// Get available discovery methods for the current system.
-///
-/// All remaining methods are userspace and unconditionally available.
-pub fn get_available_methods() -> Vec<(DiscoveryMethod, bool)> {
-    vec![
-        (DiscoveryMethod::TcpConnect, true),
-        (DiscoveryMethod::Mdns, true),
-        (DiscoveryMethod::Ssdp, true),
-        (DiscoveryMethod::UdpScan, true),
-        (DiscoveryMethod::WsDiscovery, true),
-        (DiscoveryMethod::ArpCache, true),
-        (DiscoveryMethod::Snmp, true),
-        (DiscoveryMethod::Llmnr, true),
-        (DiscoveryMethod::None, true),
-    ]
-}
-
-/// Get all local (non-loopback) IP addresses of this machine
-fn get_local_ip_addresses() -> HashSet<IpAddr> {
-    let mut local_ips = HashSet::new();
-
-    if let Ok(addrs) = if_addrs::get_if_addrs() {
-        for iface in addrs {
-            let ip = iface.addr.ip();
-            // Skip loopback addresses
-            if !ip.is_loopback() {
-                local_ips.insert(ip);
-            }
-        }
-    }
-
-    local_ips
-}
-
-// =============================================================================
-// SSDP/UPnP Discovery
-// =============================================================================
-
-/// SSDP multicast address for IPv4
-const SSDP_MULTICAST_ADDR_V4: &str = "239.255.255.250:1900";
-
-/// SSDP multicast address for IPv6 link-local (ff02::c)
-const SSDP_MULTICAST_ADDR_V6: &str = "[ff02::c]:1900";
-
-/// M-SEARCH search targets for broader device discovery
-const SSDP_SEARCH_TARGETS: &[&str] = &[
-    "ssdp:all",
-    "upnp:rootdevice",
-    "urn:schemas-upnp-org:device:InternetGatewayDevice:1",
-];
-
-/// Build M-SEARCH request for given host and search target
-fn build_msearch(host: &str, st: &str) -> Vec<u8> {
-    format!(
-        "M-SEARCH * HTTP/1.1\r\n\
-         Host: {host}\r\n\
-         Man: \"ssdp:discover\"\r\n\
-         MX: 2\r\n\
-         ST: {st}\r\n\
-         \r\n"
-    )
-    .into_bytes()
-}
-
-/// Per-host SSDP data collected during discovery
-#[derive(Clone, Default)]
-struct SsdpHostData {
-    server: Option<String>,
-    location: Option<String>,
-}
-
-/// Minimum timeout for SSDP discovery (multicast needs time for responses)
-const SSDP_MIN_TIMEOUT_MS: u32 = 3000;
-
-/// Maximum concurrent LOCATION XML fetches
-const SSDP_MAX_LOCATION_FETCHES: usize = 5;
-
-/// Discover hosts using SSDP/UPnP protocol with device description fetching
-async fn ssdp_discovery(options: &DiscoveryOptions) -> DiscoveryResult {
-    let start = std::time::Instant::now();
-    let ssdp_timeout = options.timeout_ms.max(SSDP_MIN_TIMEOUT_MS);
-    let timeout_duration = Duration::from_millis(u64::from(ssdp_timeout));
-
-    let hosts = Arc::new(tokio::sync::Mutex::new(HashSet::new()));
-    let host_data = Arc::new(tokio::sync::Mutex::new(std::collections::HashMap::<
-        String,
-        SsdpHostData,
-    >::new()));
-
-    // Run IPv4 and IPv6 SSDP discovery in parallel
-    let v4_handle = {
-        let hosts = Arc::clone(&hosts);
-        let host_data = Arc::clone(&host_data);
-        tokio::spawn(async move {
-            ssdp_collect_responses(
-                "0.0.0.0:0",
-                SSDP_MULTICAST_ADDR_V4,
-                "239.255.255.250:1900",
-                timeout_duration,
-                hosts,
-                host_data,
-            )
-            .await;
-        })
-    };
-
-    let v6_handle = {
-        let hosts = Arc::clone(&hosts);
-        let host_data = Arc::clone(&host_data);
-        tokio::spawn(async move {
-            ssdp_collect_responses(
-                "[::]:0",
-                SSDP_MULTICAST_ADDR_V6,
-                "[ff02::c]:1900",
-                timeout_duration,
-                hosts,
-                host_data,
-            )
-            .await;
-        })
-    };
-
-    let _ = tokio::join!(v4_handle, v6_handle);
-
-    // Snapshot data to avoid holding locks during async fetch
-    let hosts_snapshot = hosts.lock().await.clone();
-    let host_data_snapshot = host_data.lock().await.clone();
-    let host_metadata = ssdp_fetch_device_descriptions(&hosts_snapshot, &host_data_snapshot).await;
-
-    DiscoveryResult {
-        method: "ssdp".to_string(),
-        hosts: hosts_snapshot.iter().cloned().collect(),
-        hostnames: std::collections::HashMap::new(),
-        host_metadata,
-        unreachable: vec![],
-        duration_ms: start.elapsed().as_millis() as u64,
-        error: None,
-        requires_privileges: false,
-    }
-}
-
-/// Fetch UPnP device description XMLs and build host metadata
-async fn ssdp_fetch_device_descriptions(
-    hosts: &HashSet<String>,
-    host_data: &std::collections::HashMap<String, SsdpHostData>,
-) -> std::collections::HashMap<String, HostMetadata> {
-    let semaphore = Arc::new(Semaphore::new(SSDP_MAX_LOCATION_FETCHES));
-    let client = Arc::new(
-        reqwest::Client::builder()
-            .connect_timeout(Duration::from_secs(1))
-            .timeout(Duration::from_secs(2))
-            .redirect(reqwest::redirect::Policy::none())
-            .build()
-            .unwrap_or_default(),
-    );
-    let mut fetch_handles = Vec::new();
-
-    for (ip, data) in host_data {
-        if let Some(ref location) = data.location {
-            let ip = ip.clone();
-            let location = location.clone();
-            let server = data.server.clone();
-            let sem = Arc::clone(&semaphore);
-            let client = Arc::clone(&client);
-
-            fetch_handles.push(tokio::spawn(async move {
-                let _permit = sem.acquire().await;
-                let mut device_info = fetch_upnp_device_description(&client, &location, &ip)
-                    .await
-                    .unwrap_or_default();
-                device_info.server = server;
-                if device_info.location.is_none() {
-                    device_info.location = Some(location);
-                }
-                (ip, device_info)
-            }));
-        }
-    }
-
-    let mut metadata = std::collections::HashMap::<String, HostMetadata>::new();
-
-    for handle in fetch_handles {
-        if let Ok((ip, device_info)) = handle.await {
-            // Only use friendlyName as hostname (SERVER is not an identifier)
-            let has_friendly_name = device_info.friendly_name.is_some();
-            metadata.insert(
-                ip,
-                HostMetadata {
-                    hostname: device_info.friendly_name.clone(),
-                    hostname_source: if has_friendly_name {
-                        Some("ssdp".to_string())
-                    } else {
-                        None
-                    },
-                    ssdp_device: Some(device_info),
-                    ..Default::default()
-                },
-            );
-        }
-    }
-
-    // Add basic metadata for hosts without LOCATION or failed fetch
-    for ip in hosts {
-        if !metadata.contains_key(ip) {
-            if let Some(data) = host_data.get(ip) {
-                if data.server.is_some() || data.location.is_some() {
-                    metadata.insert(
-                        ip.clone(),
-                        HostMetadata {
-                            ssdp_device: Some(super::types::SsdpDeviceInfo {
-                                server: data.server.clone(),
-                                ..Default::default()
-                            }),
-                            ..Default::default()
-                        },
-                    );
-                }
-            }
-        }
-    }
-
-    metadata
-}
-
-/// Collect SSDP M-SEARCH responses from a specific address family
-async fn ssdp_collect_responses(
-    bind_addr: &str,
-    multicast_addr: &str,
-    host_header: &str,
-    timeout_duration: Duration,
-    hosts: Arc<tokio::sync::Mutex<HashSet<String>>>,
-    host_data: Arc<tokio::sync::Mutex<std::collections::HashMap<String, SsdpHostData>>>,
-) {
-    let socket = match tokio::net::UdpSocket::bind(bind_addr).await {
-        Ok(s) => s,
-        Err(e) => {
-            eprintln!("SSDP: Failed to bind {bind_addr}: {e}");
-            return;
-        }
-    };
-
-    // Send M-SEARCH for each search target
-    if let Ok(addr) = multicast_addr.parse::<SocketAddr>() {
-        for st in SSDP_SEARCH_TARGETS {
-            let msg = build_msearch(host_header, st);
-            let _ = socket.send_to(&msg, addr).await;
-        }
-    }
-
-    // Collect responses
-    let mut buf = [0u8; 4096];
-    let deadline = tokio::time::Instant::now() + timeout_duration;
-
-    loop {
-        let remaining = deadline.saturating_duration_since(tokio::time::Instant::now());
-        if remaining.is_zero() {
-            break;
-        }
-
-        match tokio::time::timeout(
-            remaining.min(Duration::from_millis(100)),
-            socket.recv_from(&mut buf),
-        )
-        .await
-        {
-            Ok(Ok((len, addr))) => {
-                if addr.ip().is_loopback() {
-                    continue;
-                }
-
-                let ip = addr.ip().to_string();
-                let mut hosts_guard = hosts.lock().await;
-                let is_new = hosts_guard.insert(ip.clone());
-                drop(hosts_guard);
-
-                if is_new {
-                    let data = if let Ok(response) = std::str::from_utf8(&buf[..len]) {
-                        let parsed = parse_ssdp_response(response);
-                        SsdpHostData {
-                            server: parsed.server,
-                            location: parsed.location,
-                        }
-                    } else {
-                        SsdpHostData::default()
-                    };
-                    host_data.lock().await.insert(ip, data);
-                } else if let Ok(response) = std::str::from_utf8(&buf[..len]) {
-                    // Update with better data if we get a LOCATION from a later response
-                    let parsed = parse_ssdp_response(response);
-                    if parsed.location.is_some() {
-                        let mut data_guard = host_data.lock().await;
-                        if let Some(existing) = data_guard.get_mut(&ip) {
-                            if existing.location.is_none() {
-                                existing.location = parsed.location;
-                            }
-                            if existing.server.is_none() {
-                                existing.server = parsed.server;
-                            }
-                        }
-                    }
-                }
-            }
-            Ok(Err(_)) | Err(_) => continue,
-        }
-    }
-}
-
-/// Parsed SSDP response headers
-struct SsdpResponse {
-    server: Option<String>,
-    location: Option<String>,
-}
-
-/// Parse SERVER and LOCATION headers from SSDP response
-fn parse_ssdp_response(response: &str) -> SsdpResponse {
-    let mut server = None;
-    let mut location = None;
-
-    for line in response.lines() {
-        let lower = line.to_lowercase();
-        if lower.starts_with("server:") {
-            server = Some(line[7..].trim().to_string());
-        } else if lower.starts_with("location:") {
-            location = Some(line[9..].trim().to_string());
-        }
-    }
-
-    SsdpResponse { server, location }
-}
-
-/// Maximum response body size for UPnP device description XML (64 KB)
-const UPNP_MAX_BODY_SIZE: usize = 64 * 1024;
-
-/// Validate that a LOCATION URL is safe to fetch (SSRF prevention).
-/// Checks: HTTP scheme only, host is private/link-local, host matches SSDP source IP.
-fn is_safe_upnp_location(location: &str, source_ip: &str) -> bool {
-    let url = match reqwest::Url::parse(location) {
-        Ok(u) => u,
-        Err(_) => return false,
-    };
-
-    // Only allow HTTP (UPnP descriptions are HTTP-only)
-    if url.scheme() != "http" {
-        return false;
-    }
-
-    // Must have a valid host that is an IP address
-    let host = match url.host_str() {
-        Some(h) => h,
-        None => return false,
-    };
-
-    // Strip IPv6 brackets if present (host_str() returns "[::1]" for IPv6)
-    let host_bare = host
-        .strip_prefix('[')
-        .and_then(|h| h.strip_suffix(']'))
-        .unwrap_or(host);
-
-    // Parse both host and source_ip as IpAddr for canonical comparison
-    let location_ip: IpAddr = match host_bare.parse() {
-        Ok(ip) => ip,
-        Err(_) => return false, // Non-IP hostnames are rejected
-    };
-    let source: IpAddr = match source_ip.parse() {
-        Ok(ip) => ip,
-        Err(_) => return false,
-    };
-
-    // Verify LOCATION host matches the SSDP response source IP
-    if location_ip != source {
-        return false;
-    }
-
-    // Verify it's a private/link-local address
-    match location_ip {
-        IpAddr::V4(ip) => ip.is_private() || ip.is_link_local(),
-        IpAddr::V6(ip) => {
-            // fe80::/10 link-local or fc00::/7 unique-local (RFC 4193)
-            let segments = ip.segments();
-            (segments[0] & 0xffc0) == 0xfe80 || (segments[0] & 0xfe00) == 0xfc00
-        }
-    }
-}
-
-/// Fetch and parse UPnP device description XML from LOCATION URL
-async fn fetch_upnp_device_description(
-    client: &reqwest::Client,
-    location: &str,
-    source_ip: &str,
-) -> Option<super::types::SsdpDeviceInfo> {
-    // Validate URL before fetching (SSRF prevention)
-    if !is_safe_upnp_location(location, source_ip) {
-        return None;
-    }
-
-    let response = client.get(location).send().await.ok()?;
-
-    // Check content-length before reading body (fast rejection)
-    if let Some(content_length) = response.content_length() {
-        if content_length as usize > UPNP_MAX_BODY_SIZE {
-            return None;
-        }
-    }
-
-    // Read body with streaming size limit to prevent memory exhaustion
-    // from chunked transfer encoding without Content-Length
-    let bytes = response.bytes().await.ok()?;
-    if bytes.len() > UPNP_MAX_BODY_SIZE {
-        return None;
-    }
-    let body = String::from_utf8_lossy(&bytes);
-
-    parse_upnp_device_xml(&body, location)
-}
-
-/// Parse UPnP device description XML to extract device info
-fn parse_upnp_device_xml(xml: &str, location: &str) -> Option<super::types::SsdpDeviceInfo> {
-    use quick_xml::events::Event;
-    use quick_xml::reader::Reader;
-
-    let mut reader = Reader::from_str(xml);
-    let mut buf = Vec::new();
-
-    let mut friendly_name = None;
-    let mut manufacturer = None;
-    let mut model_name = None;
-    let mut model_number = None;
-    let mut device_type = None;
-    let mut udn = None;
-    let mut current_tag = String::new();
-    // Only capture fields from the first (root) <device> block
-    let mut device_depth: u32 = 0;
-
-    let mut assign_field = |tag: &str, text: String| match tag {
-        "friendlyName" => friendly_name = Some(text),
-        "manufacturer" => manufacturer = Some(text),
-        "modelName" => model_name = Some(text),
-        "modelNumber" => model_number = Some(text),
-        "deviceType" => device_type = Some(text),
-        "UDN" => udn = Some(text),
-        _ => {}
-    };
-
-    loop {
-        match reader.read_event_into(&mut buf) {
-            Ok(Event::Start(ref e)) => {
-                let tag = String::from_utf8_lossy(e.local_name().as_ref()).to_string();
-                if tag == "device" {
-                    device_depth += 1;
-                }
-                current_tag = tag;
-            }
-            Ok(Event::Text(ref e)) => {
-                if device_depth == 1 {
-                    let text = e
-                        .xml_content(quick_xml::XmlVersion::Implicit1_0)
-                        .unwrap_or_default()
-                        .trim()
-                        .to_string();
-                    if !text.is_empty() {
-                        assign_field(&current_tag, text);
-                    }
-                }
-            }
-            Ok(Event::CData(ref e)) => {
-                if device_depth == 1 {
-                    let text = String::from_utf8_lossy(e.as_ref()).trim().to_string();
-                    if !text.is_empty() {
-                        assign_field(&current_tag, text);
-                    }
-                }
-            }
-            Ok(Event::End(ref e)) => {
-                let tag = String::from_utf8_lossy(e.local_name().as_ref()).to_string();
-                if tag == "device" {
-                    device_depth = device_depth.saturating_sub(1);
-                }
-                current_tag.clear();
-            }
-            Ok(Event::Eof) => break,
-            Err(_) => break,
-            _ => {}
-        }
-        buf.clear();
-    }
-
-    // Only return if we found at least one useful field
-    if friendly_name.is_some()
-        || manufacturer.is_some()
-        || model_name.is_some()
-        || device_type.is_some()
-        || udn.is_some()
-    {
-        Some(super::types::SsdpDeviceInfo {
-            friendly_name,
-            manufacturer,
-            model_name,
-            model_number,
-            device_type,
-            location: Some(location.to_string()),
-            server: None,
-            udn,
-        })
-    } else {
-        None
-    }
-}
-
-// =============================================================================
-// UDP Scan Discovery
-// =============================================================================
-
-/// UDP ports to probe for host discovery
-const UDP_DISCOVERY_PORTS: &[u16] = &[
-    53,   // DNS
-    67,   // DHCP Server
-    68,   // DHCP Client
-    123,  // NTP
-    137,  // NetBIOS Name
-    138,  // NetBIOS Datagram
-    161,  // SNMP
-    500,  // IKE (VPN)
-    1900, // SSDP/UPnP
-    5353, // mDNS
-];
-
-/// Get UDP probe packet for a specific port
-fn get_udp_probe(port: u16) -> &'static [u8] {
-    match port {
-        // DNS query header (minimal)
-        53 => &[
-            0x00, 0x01, // Transaction ID
-            0x01, 0x00, // Flags: Standard query
-            0x00, 0x01, // Questions: 1
-            0x00, 0x00, // Answer RRs: 0
-            0x00, 0x00, // Authority RRs: 0
-            0x00, 0x00, // Additional RRs: 0
-        ],
-        // NetBIOS name query
-        137 => &[
-            0x80, 0x94, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x43,
-            0x4b,
-        ],
-        // SNMP GetRequest (community: public)
-        161 => &[
-            0x30, 0x26, 0x02, 0x01, 0x01, 0x04, 0x06, 0x70, 0x75, 0x62, 0x6c, 0x69, 0x63,
-        ],
-        // NTP version request
-        123 => &[0x1b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00],
-        // Generic probe
-        _ => &[0x00],
-    }
-}
-
-/// Discover hosts using UDP probes to common ports
-async fn udp_scan_discovery(targets: &[IpAddr], options: &DiscoveryOptions) -> DiscoveryResult {
-    let start = std::time::Instant::now();
-    let probe_timeout = Duration::from_millis(200);
-    let semaphore = Arc::new(Semaphore::new(options.concurrency as usize));
-
-    let mut hosts = Vec::new();
-    let mut unreachable = Vec::new();
-
-    // Filter to IPv4 targets only (UDP scan implementation is simpler for IPv4)
-    let ipv4_targets: Vec<IpAddr> = targets.iter().filter(|ip| ip.is_ipv4()).copied().collect();
-
-    if ipv4_targets.is_empty() {
-        return DiscoveryResult {
-            method: "udp_scan".to_string(),
-            hosts: vec![],
-            hostnames: std::collections::HashMap::new(),
-            host_metadata: std::collections::HashMap::new(),
-            unreachable: vec![],
-            duration_ms: start.elapsed().as_millis() as u64,
-            error: None,
-            requires_privileges: false,
-        };
-    }
-
-    let mut handles = Vec::with_capacity(ipv4_targets.len());
-
-    for target in ipv4_targets {
-        let sem = Arc::clone(&semaphore);
-
-        let handle = tokio::spawn(async move {
-            let _permit = sem.acquire().await;
-
-            // Create a single socket for this target
-            let socket = match tokio::net::UdpSocket::bind("0.0.0.0:0").await {
-                Ok(s) => s,
-                Err(_) => return (target, false),
-            };
-
-            // Try UDP probes to common ports
-            for &port in UDP_DISCOVERY_PORTS {
-                let addr = SocketAddr::new(target, port);
-                let probe = get_udp_probe(port);
-
-                // Send probe
-                if socket.send_to(probe, addr).await.is_err() {
-                    continue;
-                }
-
-                // Wait for response with timeout
-                let mut buf = [0u8; 512];
-                match tokio::time::timeout(probe_timeout, socket.recv_from(&mut buf)).await {
-                    Ok(Ok(_)) => return (target, true),
-                    Ok(Err(_)) | Err(_) => continue,
-                }
-            }
-
-            (target, false)
-        });
-
-        handles.push(handle);
-    }
-
-    // Collect all results
-    for handle in handles {
-        if let Ok((target, reachable)) = handle.await {
-            if reachable {
-                hosts.push(target.to_string());
-            } else {
-                unreachable.push(target.to_string());
-            }
-        }
-    }
-
-    DiscoveryResult {
-        method: "udp_scan".to_string(),
-        hosts,
-        hostnames: std::collections::HashMap::new(),
-        host_metadata: std::collections::HashMap::new(),
-        unreachable,
-        duration_ms: start.elapsed().as_millis() as u64,
-        error: None,
-        requires_privileges: false,
-    }
-}
-
-// =============================================================================
-// ARP Cache Discovery (non-privileged)
-// =============================================================================
-
-/// Discover hosts by reading the OS ARP cache (no privileges required)
-fn arp_cache_discovery(targets: &[IpAddr]) -> DiscoveryResult {
-    let start = std::time::Instant::now();
-
-    // Read both IPv4 ARP and IPv6 NDP caches to expose MAC addresses for both families
-    let entries = super::arp_cache::read_neighbor_cache();
-    let target_set: HashSet<String> = targets.iter().map(ToString::to_string).collect();
-
-    let mut hosts = Vec::new();
-    let mut host_metadata = std::collections::HashMap::new();
-
-    for entry in &entries {
-        // Only include entries that match our target range
-        if !target_set.contains(&entry.ip) {
-            // Check if the IP is within any target CIDR
-            let entry_ip: IpAddr = match entry.ip.parse() {
-                Ok(ip) => ip,
-                Err(_) => continue,
-            };
-            if !targets.contains(&entry_ip) {
-                continue;
-            }
-        }
-
-        hosts.push(entry.ip.clone());
-
-        // Look up vendor from OUI
-        let vendor = super::oui::lookup_vendor(&entry.mac).map(String::from);
-
-        host_metadata.insert(
-            entry.ip.clone(),
-            HostMetadata {
-                mac_address: Some(entry.mac.clone()),
-                vendor,
-                ..Default::default()
-            },
-        );
-    }
-
-    DiscoveryResult {
-        method: "arp_cache".to_string(),
-        hosts,
-        hostnames: std::collections::HashMap::new(),
-        host_metadata,
-        unreachable: vec![],
-        duration_ms: start.elapsed().as_millis() as u64,
-        error: None,
-        requires_privileges: false,
-    }
-}
-
-// =============================================================================
-// WS-Discovery Method
-// =============================================================================
-
-/// Discover hosts using WS-Discovery SOAP Probe
-async fn ws_discovery_method(options: &DiscoveryOptions) -> DiscoveryResult {
-    let start = std::time::Instant::now();
-    let timeout_duration = Duration::from_millis(u64::from(options.timeout_ms.max(3000)));
-
-    let results = Box::pin(super::ws_discovery::ws_discovery_probe(timeout_duration)).await;
-
-    let hosts: Vec<String> = results.keys().cloned().collect();
-    let host_metadata: std::collections::HashMap<String, HostMetadata> = results
-        .into_iter()
-        .map(|(ip, info)| {
-            (
-                ip,
-                HostMetadata {
-                    ws_discovery: Some(info),
-                    ..Default::default()
-                },
-            )
-        })
-        .collect();
-
-    DiscoveryResult {
-        method: "ws_discovery".to_string(),
-        hosts,
-        hostnames: std::collections::HashMap::new(),
-        host_metadata,
-        unreachable: vec![],
-        duration_ms: start.elapsed().as_millis() as u64,
-        error: None,
-        requires_privileges: false,
-    }
-}
-
-// =============================================================================
-// SNMP Discovery (broadcast sysName query)
-// =============================================================================
-
-/// Maximum concurrent SNMP probes
-const SNMP_DISCOVERY_CONCURRENCY: usize = 64;
-
-/// Per-host SNMP query timeout
-const SNMP_DISCOVERY_TIMEOUT: Duration = Duration::from_secs(1);
-
-/// SNMP communities to try, in order
-const SNMP_DISCOVERY_COMMUNITIES: &[&str] = &["public", "private"];
-
-/// Discover hosts via SNMP sysName GetRequest across the target range.
-///
-/// For each target, sends an SNMPv2c `sysName.0` query in a blocking task,
-/// trying each community in [`SNMP_DISCOVERY_COMMUNITIES`]. Successful
-/// responses are recorded as reachable hosts with populated `snmp_info`.
-async fn snmp_discovery(targets: &[IpAddr], _options: &DiscoveryOptions) -> DiscoveryResult {
-    let start = std::time::Instant::now();
-    let semaphore = Arc::new(Semaphore::new(SNMP_DISCOVERY_CONCURRENCY));
-
-    let handles: Vec<_> = targets
-        .iter()
-        .copied()
-        .map(|target| {
-            let sem = Arc::clone(&semaphore);
-            tokio::spawn(async move {
-                let _permit = sem.acquire().await.ok()?;
-                let mut found = None;
-                for community in SNMP_DISCOVERY_COMMUNITIES {
-                    if let Some(info) = super::snmp::query_snmp_device_info_with_community(
-                        target,
-                        community.as_bytes(),
-                        SNMP_DISCOVERY_TIMEOUT,
-                    )
-                    .await
-                    {
-                        found = Some(info);
-                        break;
-                    }
-                }
-                found.map(|info| (target, info))
-            })
-        })
-        .collect();
-
-    let mut hosts = Vec::new();
-    let mut host_metadata: std::collections::HashMap<String, HostMetadata> =
-        std::collections::HashMap::new();
-    let mut unreachable: Vec<String> = targets.iter().map(ToString::to_string).collect();
-
-    for handle in handles {
-        if let Ok(Some((target, info))) = handle.await {
-            let ip_str = target.to_string();
-            unreachable.retain(|ip| ip != &ip_str);
-            hosts.push(ip_str.clone());
-
-            let hostname = info.sys_name.clone();
-            host_metadata.insert(
-                ip_str,
-                HostMetadata {
-                    hostname: hostname.clone(),
-                    hostname_source: hostname.as_ref().map(|_| "snmp".to_string()),
-                    snmp_info: Some(info),
-                    ..Default::default()
-                },
-            );
-        }
-    }
-
-    DiscoveryResult {
-        method: "snmp".to_string(),
-        hosts,
-        hostnames: std::collections::HashMap::new(),
-        host_metadata,
-        unreachable,
-        duration_ms: start.elapsed().as_millis() as u64,
-        error: None,
-        requires_privileges: false,
-    }
-}
-
-// =============================================================================
-// LLMNR Discovery (multicast reverse PTR queries)
-// =============================================================================
-
-/// Total time budget for LLMNR collection
-const LLMNR_DISCOVERY_TIMEOUT: Duration = Duration::from_secs(2);
-
-/// Maximum concurrent LLMNR queries
-const LLMNR_DISCOVERY_CONCURRENCY: usize = 64;
-
-/// Discover hosts via LLMNR multicast PTR resolution of every target IP.
-///
-/// LLMNR has no wildcard concept, so we follow Windows' own approach: issue
-/// reverse PTR queries (`x.in-addr.arpa` for IPv4) toward the LLMNR multicast
-/// group and treat any responding host as alive with the resolved hostname.
-/// The whole batch runs concurrently; the function returns once every spawned
-/// probe finishes or its per-query timeout elapses, capped by
-/// [`LLMNR_DISCOVERY_TIMEOUT`].
-async fn llmnr_discovery(targets: &[IpAddr], _options: &DiscoveryOptions) -> DiscoveryResult {
-    let start = std::time::Instant::now();
-    let semaphore = Arc::new(Semaphore::new(LLMNR_DISCOVERY_CONCURRENCY));
-
-    let handles: Vec<_> = targets
-        .iter()
-        .copied()
-        .map(|target| {
-            let sem = Arc::clone(&semaphore);
-            tokio::spawn(async move {
-                let _permit = sem.acquire().await.ok()?;
-                let ip_str = target.to_string();
-                let name = super::llmnr::resolve_hostname(&ip_str, LLMNR_DISCOVERY_TIMEOUT).await?;
-                Some((target, name))
-            })
-        })
-        .collect();
-
-    let mut hostnames = std::collections::HashMap::new();
-    let mut host_metadata: std::collections::HashMap<String, HostMetadata> =
-        std::collections::HashMap::new();
-    let mut hosts = Vec::new();
-    let mut unreachable: Vec<String> = targets.iter().map(ToString::to_string).collect();
-
-    for handle in handles {
-        if let Ok(Some((target, name))) = handle.await {
-            let ip_str = target.to_string();
-            unreachable.retain(|ip| ip != &ip_str);
-            hosts.push(ip_str.clone());
-            hostnames.insert(ip_str.clone(), name.clone());
-            host_metadata.insert(
-                ip_str,
-                HostMetadata {
-                    hostname: Some(name),
-                    hostname_source: Some("llmnr".to_string()),
-                    ..Default::default()
-                },
-            );
-        }
-    }
-
-    DiscoveryResult {
-        method: "llmnr".to_string(),
-        hosts,
-        hostnames,
-        host_metadata,
-        unreachable,
-        duration_ms: start.elapsed().as_millis() as u64,
-        error: None,
-        requires_privileges: false,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    mod is_safe_upnp_location {
-        use super::*;
-
-        #[test]
-        fn test_valid_private_ipv4_http() {
-            assert!(is_safe_upnp_location(
-                "http://192.168.1.1:49152/rootDesc.xml",
-                "192.168.1.1"
-            ));
-        }
-
-        #[test]
-        fn test_valid_10_network() {
-            assert!(is_safe_upnp_location(
-                "http://10.0.0.1:8080/desc.xml",
-                "10.0.0.1"
-            ));
-        }
-
-        #[test]
-        fn test_valid_172_network() {
-            assert!(is_safe_upnp_location(
-                "http://172.16.0.1:1234/desc.xml",
-                "172.16.0.1"
-            ));
-        }
-
-        #[test]
-        fn test_rejects_https() {
-            assert!(!is_safe_upnp_location(
-                "https://192.168.1.1:443/desc.xml",
-                "192.168.1.1"
-            ));
-        }
-
-        #[test]
-        fn test_rejects_hostname() {
-            assert!(!is_safe_upnp_location(
-                "http://mydevice.local:8080/desc.xml",
-                "192.168.1.1"
-            ));
-        }
-
-        #[test]
-        fn test_rejects_source_ip_mismatch() {
-            assert!(!is_safe_upnp_location(
-                "http://192.168.1.1:49152/rootDesc.xml",
-                "192.168.1.2"
-            ));
-        }
-
-        #[test]
-        fn test_rejects_public_ip() {
-            assert!(!is_safe_upnp_location(
-                "http://8.8.8.8:80/desc.xml",
-                "8.8.8.8"
-            ));
-        }
-
-        #[test]
-        fn test_rejects_loopback() {
-            assert!(!is_safe_upnp_location(
-                "http://127.0.0.1:80/desc.xml",
-                "127.0.0.1"
-            ));
-        }
-
-        #[test]
-        fn test_valid_ipv6_link_local() {
-            assert!(is_safe_upnp_location(
-                "http://[fe80::1]:49152/desc.xml",
-                "fe80::1"
-            ));
-        }
-
-        #[test]
-        fn test_valid_ipv6_ula_fd() {
-            assert!(is_safe_upnp_location(
-                "http://[fd00::1]:49152/desc.xml",
-                "fd00::1"
-            ));
-        }
-
-        #[test]
-        fn test_valid_ipv6_ula_fc() {
-            assert!(is_safe_upnp_location(
-                "http://[fc00::1]:49152/desc.xml",
-                "fc00::1"
-            ));
-        }
-
-        #[test]
-        fn test_rejects_ipv6_global() {
-            assert!(!is_safe_upnp_location(
-                "http://[2001:db8::1]:80/desc.xml",
-                "2001:db8::1"
-            ));
-        }
-
-        #[test]
-        fn test_rejects_empty_url() {
-            assert!(!is_safe_upnp_location("", "192.168.1.1"));
-        }
-
-        #[test]
-        fn test_rejects_invalid_url() {
-            assert!(!is_safe_upnp_location("not-a-url", "192.168.1.1"));
-        }
-    }
-
-    mod parse_upnp_device_xml {
-        use super::*;
-
-        #[test]
-        fn test_valid_device_description() {
-            let xml = r#"<?xml version="1.0"?>
-<root xmlns="urn:schemas-upnp-org:device-1-0">
-  <device>
-    <friendlyName>My Router</friendlyName>
-    <manufacturer>ACME Corp</manufacturer>
-    <modelName>Router 3000</modelName>
-    <modelNumber>R3K</modelNumber>
-    <deviceType>urn:schemas-upnp-org:device:InternetGatewayDevice:1</deviceType>
-  </device>
-</root>"#;
-            let result = parse_upnp_device_xml(xml, "http://192.168.1.1:49152/desc.xml");
-            assert!(result.is_some());
-            let info = result.unwrap();
-            assert_eq!(info.friendly_name.as_deref(), Some("My Router"));
-            assert_eq!(info.manufacturer.as_deref(), Some("ACME Corp"));
-            assert_eq!(info.model_name.as_deref(), Some("Router 3000"));
-            assert_eq!(info.model_number.as_deref(), Some("R3K"));
-            assert_eq!(
-                info.device_type.as_deref(),
-                Some("urn:schemas-upnp-org:device:InternetGatewayDevice:1")
-            );
-            assert_eq!(
-                info.location.as_deref(),
-                Some("http://192.168.1.1:49152/desc.xml")
-            );
-        }
-
-        #[test]
-        fn test_empty_xml() {
-            let result = parse_upnp_device_xml("", "http://192.168.1.1/desc.xml");
-            assert!(result.is_none());
-        }
-
-        #[test]
-        fn test_xml_without_device_fields() {
-            let xml = r#"<?xml version="1.0"?><root><device></device></root>"#;
-            let result = parse_upnp_device_xml(xml, "http://192.168.1.1/desc.xml");
-            assert!(result.is_none());
-        }
-
-        #[test]
-        fn test_cdata_in_friendly_name() {
-            let xml = r#"<?xml version="1.0"?>
-<root>
-  <device>
-    <friendlyName><![CDATA[My Device & More]]></friendlyName>
-    <manufacturer>TestCo</manufacturer>
-  </device>
-</root>"#;
-            let result = parse_upnp_device_xml(xml, "http://192.168.1.1/desc.xml");
-            assert!(result.is_some());
-            let info = result.unwrap();
-            assert_eq!(info.friendly_name.as_deref(), Some("My Device & More"));
-        }
-
-        #[test]
-        fn test_nested_device_uses_root_only() {
-            let xml = r#"<?xml version="1.0"?>
-<root>
-  <device>
-    <friendlyName>Root Device</friendlyName>
-    <deviceList>
-      <device>
-        <friendlyName>Sub Device</friendlyName>
-      </device>
-    </deviceList>
-  </device>
-</root>"#;
-            let result = parse_upnp_device_xml(xml, "http://192.168.1.1/desc.xml");
-            assert!(result.is_some());
-            let info = result.unwrap();
-            assert_eq!(info.friendly_name.as_deref(), Some("Root Device"));
-        }
-
-        #[test]
-        fn test_malformed_xml() {
-            let xml = r"<root><device><friendlyName>Broken";
-            let result = parse_upnp_device_xml(xml, "http://192.168.1.1/desc.xml");
-            // Should not panic; may return None or partial result
-            // The parser breaks on Err but may have captured partial data
-            assert!(result.is_none() || result.is_some());
-        }
-
-        #[test]
-        fn test_partial_fields() {
-            let xml = r#"<?xml version="1.0"?>
-<root>
-  <device>
-    <manufacturer>OnlyManufacturer</manufacturer>
-  </device>
-</root>"#;
-            let result = parse_upnp_device_xml(xml, "http://192.168.1.1/desc.xml");
-            assert!(result.is_some());
-            let info = result.unwrap();
-            assert_eq!(info.manufacturer.as_deref(), Some("OnlyManufacturer"));
-            assert!(info.friendly_name.is_none());
-            assert!(info.model_name.is_none());
-        }
-
-        #[test]
-        fn test_extracts_udn() {
-            let xml = r#"<?xml version="1.0"?>
-<root xmlns="urn:schemas-upnp-org:device-1-0">
-  <device>
-    <friendlyName>My Router</friendlyName>
-    <UDN>uuid:8b8bcfe9-7b66-4d96-bb1e-9c9b1c5e8c3a</UDN>
-  </device>
-</root>"#;
-            let result = parse_upnp_device_xml(xml, "http://192.168.1.1/desc.xml");
-            assert!(result.is_some());
-            let info = result.unwrap();
-            assert_eq!(
-                info.udn.as_deref(),
-                Some("uuid:8b8bcfe9-7b66-4d96-bb1e-9c9b1c5e8c3a")
-            );
-        }
-
-        #[test]
-        fn test_missing_udn_returns_none() {
-            let xml = r#"<?xml version="1.0"?>
-<root xmlns="urn:schemas-upnp-org:device-1-0">
-  <device>
-    <friendlyName>My Router</friendlyName>
-  </device>
-</root>"#;
-            let result = parse_upnp_device_xml(xml, "http://192.168.1.1/desc.xml");
-            assert!(result.is_some());
-            let info = result.unwrap();
-            assert!(info.udn.is_none());
-        }
-
-        #[test]
-        fn test_udn_alone_triggers_some() {
-            // Even without other fields, a UDN-only document should still
-            // produce SsdpDeviceInfo because UDN is a useful identity signal.
-            let xml = r#"<?xml version="1.0"?>
-<root xmlns="urn:schemas-upnp-org:device-1-0">
-  <device>
-    <UDN>uuid:11111111-2222-3333-4444-555555555555</UDN>
-  </device>
-</root>"#;
-            let result = parse_upnp_device_xml(xml, "http://192.168.1.1/desc.xml");
-            assert!(result.is_some());
-            let info = result.unwrap();
-            assert_eq!(
-                info.udn.as_deref(),
-                Some("uuid:11111111-2222-3333-4444-555555555555")
-            );
-        }
-    }
-
-    mod parse_ssdp_response {
-        use super::*;
-
-        #[test]
-        fn test_with_server_and_location() {
-            let response = "HTTP/1.1 200 OK\r\n\
-                SERVER: Linux/3.0 UPnP/1.0 miniupnpd/1.5\r\n\
-                LOCATION: http://192.168.1.1:49152/rootDesc.xml\r\n\
-                ST: upnp:rootdevice\r\n\r\n";
-            let parsed = parse_ssdp_response(response);
-            assert_eq!(
-                parsed.server.as_deref(),
-                Some("Linux/3.0 UPnP/1.0 miniupnpd/1.5")
-            );
-            assert_eq!(
-                parsed.location.as_deref(),
-                Some("http://192.168.1.1:49152/rootDesc.xml")
-            );
-        }
-
-        #[test]
-        fn test_without_location() {
-            let response = "HTTP/1.1 200 OK\r\n\
-                SERVER: Some Server\r\n\
-                ST: ssdp:all\r\n\r\n";
-            let parsed = parse_ssdp_response(response);
-            assert_eq!(parsed.server.as_deref(), Some("Some Server"));
-            assert!(parsed.location.is_none());
-        }
-
-        #[test]
-        fn test_empty_response() {
-            let parsed = parse_ssdp_response("");
-            assert!(parsed.server.is_none());
-            assert!(parsed.location.is_none());
-        }
-
-        #[test]
-        fn test_case_insensitive_headers() {
-            let response = "HTTP/1.1 200 OK\r\n\
-                server: my server\r\n\
-                location: http://10.0.0.1/desc.xml\r\n\r\n";
-            let parsed = parse_ssdp_response(response);
-            assert_eq!(parsed.server.as_deref(), Some("my server"));
-            assert_eq!(parsed.location.as_deref(), Some("http://10.0.0.1/desc.xml"));
-        }
-    }
-}
